modifer cargasonServiceImpl

+++ src/main/java/org/fleetflow/souktransportbackend/service/impl/CargaisonServiceImpl.java
@@ -64,7 +64,9 @@ public class CargaisonServiceImpl implements CargaisonService {
         if (dto.getPoids() != null && dto.getPoids() <= 0) {
             throw new IllegalArgumentException("Le poids doit être supérieur à 0");
         }
-
+        if (cargaison.getStatutCargaison() != StatutCargaison.SOUMISE) {
+            throw new IllegalStateException("Impossible de modifier une cargaison qui est en transit, livrée ou annulée.");
+        }
         if (dto.getExpediteurId() != null) {
             Expediteur expediteur = expediteurRepository.findById(dto.getExpediteurId()).orElseThrow(() -> new EntityNotFoundException("Expéditeur introuvable avec l'id : " + dto.getExpediteurId()));
             cargaison.setExpediteur(expediteur);
@@ -77,6 +79,9 @@ public class CargaisonServiceImpl implements CargaisonService {
     @Override
     public void supprimerCargaison(Long id) {
         Cargaison cargaison = cargaisonRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Cargaison introuvable avec l'id : " + id));
+        if (cargaison.getStatutCargaison() != StatutCargaison.SOUMISE) {
+            throw new IllegalStateException("Impossible de supprimer une cargaison qui est en transit ou livrée.");
+        }
         cargaisonRepository.delete(cargaison);
     }
